+++ examples/deploy_voluptuous.py
@@ -2,6 +2,7 @@
 
 import nestedtext as nt
 from voluptuous import Schema, Coerce, MultipleInvalid
+from voluptuous_errors import report_voluptuous_errors
 from inform import error, full_stop, terminate
 from pprint import pprint
 
@@ -18,33 +19,18 @@ schema = Schema({
     'webmaster_email': str,
 })
 
+def normalize_key(key, parent_keys):
+    return '_'.join(key.lower().split())
+
 filename = "deploy.nt"
 try:
     keymap = {}
-    raw = nt.load(filename, keymap=keymap)
+    raw = nt.load(filename, keymap=keymap, normalize_key=normalize_key)
     config = schema(raw)
 except nt.NestedTextError as e:
     e.terminate()
-except MultipleInvalid as exception:
-    voluptuous_error_messages = {  # provide user-friendly error messages
-        "extra keys not allowed": ("unknown key", "key"),
-        "expected a dictionary": ("expected key-value pair", "value"),
-    }
-    for e in exception.errors:
-        msg, flag = voluptuous_error_messages.get(
-            e.msg, (e.msg, 'value')
-        )
-        culprit = nt.join_keys(e.path, keymap=keymap, sep="/")
-        loc = keymap.get(tuple(e.path))
-        if loc:
-            codicil = loc.as_line(flag)
-            line_num, col_num = loc.as_tuple(flag)
-            source = f"{filename!s}@{line_num+1}"
-        else:  # required key is missing
-            codicil = None
-            source = str(filename)
-
-        error(full_stop(msg), culprit=(source, culprit), codicil=codicil)
+except MultipleInvalid as e:
+    report_voluptuous_errors(e, keymap, filename)
     terminate()
 
 pprint(config)